[Patchset 8] WIP Add JS<>WASM TraceProcessor
Still work in progress.

+++ src/trace_processor/sched.cc
@@ -40,21 +40,21 @@ BlobReader* g_reader;
 
 void OnReadComplete(uint32_t off, const uint8_t*, size_t len) {
   auto now = base::GetWallTimeMs();
-  static std::vector<int64_t>* lat = new std::vector<int64_t>;
-  lat->push_back((now - t_last_req).count());
+  static std::vector<int>* lat = new std::vector<int>;
+  lat->push_back(static_cast<int>((now - t_last_req).count()));
 
   static uint32_t bytes_read = 0;
   static uint32_t blocks_read = 0;
   PERFETTO_DCHECK(off == bytes_read);
   bytes_read += len;
   if (++blocks_read >= kNumBlocks) {
-    int64_t ms = (now - t_start).count();
+    int ms = static_cast<int>((now - t_start).count());
 
-    PERFETTO_ILOG("ReadCompelte %u KB (%u blocks) in %lld ms. %.2f KB/s",
+    PERFETTO_ILOG("ReadCompelte %u KB (%u blocks) in %d ms. %.2f KB/s",
                   bytes_read / 1024, blocks_read, ms,
                   bytes_read * 1000.0 / ms / 1024);
     PERFETTO_ILOG(
-        "Latency RTT (C++ -> JS -> C++) [ms.]: min: %lld, max: %lld, avg: "
+        "Latency RTT (C++ -> JS -> C++) [ms.]: min: %d, max: %d, avg: "
         "%.3lf",
         *std::min_element(lat->begin(), lat->end()),
         *std::max_element(lat->begin(), lat->end()),
